Produce Ballot Spreadsheet - Add Email Occurrance Count - #28
Add an email occurance count to the exported data.

Most commercial mailing lists (mailchimp, phplists) prevent sending more than one email to the same address within one mailshot. This is a problem where partners or families
are sharing the same email address. For the ballot we need to be able to send a separate token for each individual so therefore need to send a different emails to the
same email address within the same mailshot.

The solution is to put the second occurrance of an email address in a second mailshot and so on.

An email occurrence count was added to each record to allow the records to be split into their respective mailshots.