Produce Ballot Spreadsheet - Add CCB & RCC. Use ID Name - #28
The Constituent Caving Bodies & Regional Caving Councils have been added to the database and need to be included in the ballot.

- Added CCB & RCC to the ballot mailing list selection criteria.

The id_name of the user is either the full_name for individual members or the organisation for group members. With that addition of group members to the database, use id_name in preference to full_name.
- Updated the mailing lists to use id_name rather than full_name.
- In the ballot list the id_name is added rather than replacing full_name since both are useful when manipulating the list in a spreadsheet.