Produce Ballot Spreadsheet - Concatenate Repeated Emails - #28
Concatenate the name and ballot token fields of multiple occurences of the same email address to form a single email record.

Most commercial mailing lists (mailerlite, mailchimp, phplists) prevent sending more than one email to the same address. This is a problem where partners or families are sharing the same email address. The solution is to concatenate the name and ballot token fields of multiple occurrences of the same email address to form a single email record. The downside is that one person might steal the ballot tokens of the partner/family but that is unlikely. If it is an issue they shouldn't share email addresses.

The previous approach of adding an occurence count to the repeated records was insufficient. Although it is worth retaining the email occurrence count to easily identify those records that have been concatenated.

- The id_name, full_name and ballot_id fields are now concatenated.
- Use case-insensitive comparison for sort order and email comparison in the concatenation routine.
  - Emails that only differed in case were not being concatenated.
- Fixed some more menu buttons so they appear or not depending upon user permissions.